:recycle:  Reorganize timeline folder tree structure
