+++ src/views/pages/projectDetailsPage/components/timeline/components/ContentArea.tsx
