+++ src/views/pages/projectDetailsPage/components/timeline/helpers/index.ts
@@ -1,0 +1,1 @@
+export * from './mapTimelineItemList';
