+++ src/views/pages/projectDetailsPage/components/timeline/helpers/mapTimelineItemList.ts
