+++ src/views/pages/projectDetailsPage/components/timeline/index.ts
@@ -1,9 +1,5 @@
-export * from './ContentArea'
-export * from './Date'
-export * from './DesignationItem'
-export * from './ImportItem'
-export * from './ItemTitle'
-export * from './mapTimelineItemList'
-export * from './stateIcons'
-export * from './Timeline'
-export * from './TimelineItem'
+export * from './components';
+export * from './helpers';
+export * from './DesignationItem';
+export * from './ImportItem';
+export * from './Timeline';
